requirements.txt
pybitcointools
mnemonic
datadiff
